chore: Rename serializer mixin to json-api-serializer mixin
